moved string functions to seperate header
